Reorganize to keep the top level package clean
